+++ enigma/alphabet.go
@@ -1,5 +1,10 @@
 package enigma
 
+import (
+	"math/rand/v2"
+	"strings"
+)
+
 const ALPHABET = "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
 
 func LetterToIdx(letter rune) int {
@@ -10,4 +15,15 @@ func LetterToIdx(letter rune) int {
 func IdxToLetter(idx int) rune {
 	letter := rune(idx)%26 + 'A'
 	return letter
+}
+
+func RandomLetters(count int) string {
+	var letters []string = make([]string, count)
+
+	for i := range count {
+		idx := rand.IntN(len(ALPHABET))
+		letters[i] = string(ALPHABET[idx])
+	}
+
+	return strings.Join(letters, "")
 }