+++ enigma/key_test.go
@@ -13,10 +13,9 @@ var _ = Describe("Key", func() {
 
 		Expect(err).To(BeNil())
 
-		Expect(key.ReflId).To(Equal("B"))
-		Expect(key.RotorIds).To(Equal([]string{"I", "II", "III"}))
-		Expect(key.RotorCfg).To(Equal("AAA"))
-		Expect(key.PbSpec).To(Equal([]string{"AZ", "XY", "EM"}))
+		Expect(key.Reflector).To(Equal('B'))
+		Expect(key.Rotors).To(Equal([]string{"I", "II", "III"}))
+		Expect(key.PlugboardSwaps).To(Equal([]string{"AZ", "XY", "EM"}))
 	})
 
 	It("Fails on invalid Key Spec: Empty", func() {
@@ -41,13 +40,5 @@ var _ = Describe("Key", func() {
 		Expect(key).To(BeNil())
 		Expect(err).To(Not(BeNil()))
 		Expect(err).To(MatchError("Invalid Rotor Specs: [I,II]"))
-	})
-
-	It("Fails on invalid Key Spec: Bad Rotor Format", func() {
-		key, err := enigma.ParseKeySpec("B:I@B,II@A,@D")
-
-		Expect(key).To(BeNil())
-		Expect(err).To(Not(BeNil()))
-		Expect(err).To(MatchError("Invalid Rotor: [@D]"))
 	})
 })