+++ CHANGELOG.md
@@ -1,8 +1,18 @@
 # CHANGELOG
 
+## 2026.08.16
+* Use `rune` for Reflector ID instead of `string`
+* Updated the `enigma.Key` struct to be more like the `keysheet.Entry` struct, then reuse `Key` in `Entry` ... Dedupes a bit of code.
+* Initial integration of Key Sheet into `encipher` command
+* Bit of code refactoring
+
+## 2026.08.15
+* `key-sheet` -> `keysheet`
+* `keysheet generate` now saves the generated keysheet to a file
+* Added `keysheet view` command
+
 ## 2026.08.09
 * Added `key-sheet generate` command
-* 
 
 ## 2026.08.03
 * Added more of error handling